requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6
pandas==2.2.0
numpy==1.26.3
prophet==1.1.5
pydantic-settings==2.1.0
httpx
python-docx
pypdf
openpyxl
openai>=1.10.0
email-validator==2.1.0
python-dotenv==1.0.1
